Task-39570 Allow to delete upcoming events of an occurrence

+++ agenda-webapps/src/main/webapp/vue-app/agenda-common/js/EventService.js
@@ -249,19 +249,21 @@ export function saveEventReminders(eventId, occurrenceId, reminders) {
 }
 
 export function deleteEvent(eventId) {
-  const deleteAllWebConferencesPromises = getDeleteAllWebConferencesPromises(event);
-
-  return Promise.all(deleteAllWebConferencesPromises)
-    .then(() =>
-      fetch(`${eXo.env.portal.context}/${eXo.env.portal.rest}/v1/agenda/events/${eventId}`, {
-        method: 'DELETE',
-        credentials: 'include',
-      })
-    ).then((resp) => {
+  return fetch(`${eXo.env.portal.context}/${eXo.env.portal.rest}/v1/agenda/events/${eventId}`, {
+    method: 'DELETE',
+    credentials: 'include',
+  })
+    .then((resp) => {
       if (resp && resp.ok) {
         return resp.json();
       } else {
         throw new Error('Error deleting event');
+      }
+    })
+    .then(event => {
+      if (event) {
+        const deleteAllWebConferencesPromises = getDeleteAllWebConferencesPromises(event);
+        return Promise.all(deleteAllWebConferencesPromises);
       }
     });
 }
